move legal info to about app
