+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/aboutapp/InformationDestination.kt
@@ -5,12 +5,15 @@ import android.content.Intent
 import android.os.Build
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -20,13 +23,19 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.hedvig.android.compose.ui.EmptyContentDescription
 import com.hedvig.android.core.buildconstants.HedvigBuildConstants
+import com.hedvig.android.core.icons.HedvigIcons
+import com.hedvig.android.core.icons.hedvig.small.hedvig.ArrowNorthEast
+import com.hedvig.android.core.locale.CommonLocale
 import com.hedvig.android.design.system.hedvig.ButtonDefaults.ButtonStyle.Secondary
+import com.hedvig.android.design.system.hedvig.DividerPosition
 import com.hedvig.android.design.system.hedvig.HedvigAlertDialog
 import com.hedvig.android.design.system.hedvig.HedvigButton
 import com.hedvig.android.design.system.hedvig.HedvigFullScreenCenterAlignedProgressDebounced
@@ -34,15 +43,25 @@ import com.hedvig.android.design.system.hedvig.HedvigPreview
 import com.hedvig.android.design.system.hedvig.HedvigScaffold
 import com.hedvig.android.design.system.hedvig.HedvigText
 import com.hedvig.android.design.system.hedvig.HedvigTheme
+import com.hedvig.android.design.system.hedvig.HighlightLabel
+import com.hedvig.android.design.system.hedvig.HighlightLabelDefaults
 import com.hedvig.android.design.system.hedvig.HorizontalItemsWithMaximumSpaceTaken
+import com.hedvig.android.design.system.hedvig.Icon
 import com.hedvig.android.design.system.hedvig.Surface
 import com.hedvig.android.design.system.hedvig.clearFocusOnTap
+import com.hedvig.android.design.system.hedvig.horizontalDivider
 import com.hedvig.android.design.system.hedvig.withHedvigLink
+import com.hedvig.android.language.Language
+import com.hedvig.android.language.LanguageService
 import hedvig.resources.HC_CHAT_BUTTON
+import hedvig.resources.LEGAL_A11Y
+import hedvig.resources.LEGAL_INFORMATION
+import hedvig.resources.LEGAL_PRIVACY_POLICY
 import hedvig.resources.PROFILE_ABOUT_APP_LICENSE_ATTRIBUTIONS
 import hedvig.resources.PROFILE_ABOUT_APP_MEMBER_ID
 import hedvig.resources.PROFILE_ABOUT_APP_TITLE
 import hedvig.resources.PROFILE_ABOUT_APP_VERSION
+import hedvig.resources.PROFILE_INFO_LABEL
 import hedvig.resources.Res
 import hedvig.resources.app_info_submit_bug_button
 import hedvig.resources.app_info_submit_bug_continue
@@ -55,15 +74,17 @@ import hedvig.resources.app_info_submit_bug_warning_with_chat_link_3
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun AboutAppDestination(
+internal fun InformationDestination(
   viewModel: AboutAppViewModel,
   onBackPressed: () -> Unit,
   showOpenSourceLicenses: () -> Unit,
   navigateToNewConversation: () -> Unit,
   hedvigBuildConstants: HedvigBuildConstants,
+  languageService: LanguageService,
+  openUrl: (String) -> Unit,
 ) {
   val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-  AboutAppScreen(
+  InformationScreen(
     uiState = uiState,
     onBackPressed = onBackPressed,
     showOpenSourceLicenses = showOpenSourceLicenses,
@@ -71,11 +92,13 @@ internal fun AboutAppDestination(
     appVersionName = hedvigBuildConstants.appVersionName,
     appVersionCode = hedvigBuildConstants.appVersionCode,
     navigateToNewConversation = navigateToNewConversation,
+    openUrl = openUrl,
+    languageService = languageService,
   )
 }
 
 @Composable
-private fun AboutAppScreen(
+private fun InformationScreen(
   uiState: AboutAppUiState,
   isProduction: Boolean,
   appVersionName: String,
@@ -83,9 +106,11 @@ private fun AboutAppScreen(
   onBackPressed: () -> Unit,
   showOpenSourceLicenses: () -> Unit,
   navigateToNewConversation: () -> Unit,
+  openUrl: (String) -> Unit,
+  languageService: LanguageService,
 ) {
   HedvigScaffold(
-    topAppBarText = stringResource(Res.string.PROFILE_ABOUT_APP_TITLE),
+    topAppBarText = stringResource(Res.string.PROFILE_INFO_LABEL),
     navigateUp = onBackPressed,
     modifier = Modifier.clearFocusOnTap(),
   ) {
@@ -99,7 +124,7 @@ private fun AboutAppScreen(
       }
 
       is AboutAppUiState.Content -> {
-        AboutAppContent(
+        InformationContent(
           memberId = uiState.memberId,
           deviceId = uiState.deviceId,
           showOpenSourceLicenses = showOpenSourceLicenses,
@@ -107,6 +132,8 @@ private fun AboutAppScreen(
           appVersionName = appVersionName,
           appVersionCode = appVersionCode,
           navigateToNewConversation = navigateToNewConversation,
+          openUrl = openUrl,
+          languageService = languageService,
         )
       }
     }
@@ -114,7 +141,7 @@ private fun AboutAppScreen(
 }
 
 @Composable
-private fun ColumnScope.AboutAppContent(
+private fun ColumnScope.InformationContent(
   memberId: String?,
   deviceId: String?,
   showOpenSourceLicenses: () -> Unit,
@@ -122,7 +149,15 @@ private fun ColumnScope.AboutAppContent(
   appVersionName: String,
   appVersionCode: String,
   navigateToNewConversation: () -> Unit,
+  openUrl: (String) -> Unit,
+  languageService: LanguageService,
 ) {
+  Spacer(Modifier.height(16.dp))
+  LegalInfoSection(
+    openUrl = openUrl,
+    languageService = languageService,
+  )
+  Spacer(Modifier.height(24.dp))
   var showSubmitBugWarning by remember { mutableStateOf(false) }
   if (showSubmitBugWarning) {
     SubmitBugWarningDialog(
@@ -135,13 +170,22 @@ private fun ColumnScope.AboutAppContent(
       },
     )
   }
-  Spacer(Modifier.height(16.dp))
+  HighlightLabel(
+    stringResource(Res.string.PROFILE_ABOUT_APP_TITLE),
+    size = HighlightLabelDefaults.HighLightSize.Medium,
+    color = HighlightLabelDefaults.HighlightColor.Blue(
+      HighlightLabelDefaults.HighlightShade.LIGHT,
+    ),
+    modifier = Modifier.padding(horizontal = 16.dp),
+  )
   if (memberId != null) {
     HorizontalItemsWithMaximumSpaceTaken(
       spaceBetween = 8.dp,
-      modifier = Modifier
-        .fillMaxWidth()
-        .padding(16.dp),
+      modifier = horizontalDividerModifier.then(
+        Modifier
+          .fillMaxWidth()
+          .padding(16.dp),
+      ),
       startSlot = {
         HedvigText(stringResource(Res.string.PROFILE_ABOUT_APP_MEMBER_ID))
       },
@@ -155,9 +199,11 @@ private fun ColumnScope.AboutAppContent(
     )
   }
   HorizontalItemsWithMaximumSpaceTaken(
-    modifier = Modifier
-      .fillMaxWidth()
-      .padding(16.dp),
+    modifier = horizontalDividerModifier.then(
+      Modifier
+        .fillMaxWidth()
+        .padding(16.dp),
+    ),
     startSlot = {
       HedvigText(stringResource(Res.string.PROFILE_ABOUT_APP_VERSION))
     },
@@ -268,12 +314,126 @@ private fun openEmailClientWithPrefilledData(
   )
 }
 
+private val horizontalDividerModifier = Modifier.horizontalDivider(
+  DividerPosition.Bottom, horizontalPadding = 16.dp,
+)
+
+@Composable
+internal fun LegalInfoSection(
+  openUrl: (String) -> Unit,
+  languageService: LanguageService,
+  modifier: Modifier = Modifier,
+) {
+  Column(modifier) {
+    val linkContainer = LinkContainer(languageService)
+
+    LinkRow(
+      modifier = horizontalDividerModifier,
+      text = stringResource(Res.string.LEGAL_PRIVACY_POLICY),
+      link = linkContainer.getPrivacyPolicyLink(),
+      onLinkClick = openUrl,
+    )
+    LinkRow(
+      modifier = horizontalDividerModifier,
+      text = stringResource(Res.string.LEGAL_INFORMATION),
+      link = linkContainer.getLegalInfoLink(),
+      onLinkClick = openUrl,
+    )
+    LinkRow(
+      text = stringResource(Res.string.LEGAL_A11Y),
+      link = linkContainer.getA11yLink(),
+      onLinkClick = openUrl,
+    )
+  }
+}
+
+private class LinkContainer(
+  val languageService: LanguageService,
+) {
+  private val privacyPolicyLinkEn = "https://www.hedvig.com/se-en/hedvig/privacy-policy"
+  private val privacyPolicyLinkSe = "https://www.hedvig.com/se/hedvig/personuppgifter"
+  fun getPrivacyPolicyLink(): String {
+    return when (languageService.getLanguage()) {
+      Language.SV_SE -> privacyPolicyLinkSe
+      Language.EN_SE -> privacyPolicyLinkEn
+    }
+  }
+
+  private val legalInfoLinkEn = "https://www.hedvig.com/se-en/hedvig/legal"
+  private val legalInfoLinkSe = "https://www.hedvig.com/se/hedvig/legal"
+  fun getLegalInfoLink(): String {
+    return when (languageService.getLanguage()) {
+      Language.SV_SE -> legalInfoLinkSe
+      Language.EN_SE -> legalInfoLinkEn
+    }
+  }
+
+  private val a11yLinkEn = "https://www.hedvig.com/se-en/help/accessibility"
+  private val a11yLinkSe = "https://www.hedvig.com/se/hjalp/tillganglighet"
+  fun getA11yLink(): String {
+    return when (languageService.getLanguage()) {
+      Language.SV_SE -> a11yLinkSe
+      Language.EN_SE -> a11yLinkEn
+    }
+  }
+}
+
+
+@Composable
+private fun LinkRow(
+  link: String,
+  text: String,
+  onLinkClick: (String) -> Unit,
+  modifier: Modifier = Modifier,
+) {
+  Row(
+    verticalAlignment = Alignment.CenterVertically,
+    modifier = modifier
+      .fillMaxWidth()
+      .clickable(
+        onClick = {
+          onLinkClick(link)
+        },
+        enabled = true,
+      )
+      .padding(16.dp)
+      .semantics(mergeDescendants = true) {},
+  ) {
+    HedvigText(
+      text = text,
+    )
+    Spacer(Modifier.weight(1f))
+    Spacer(Modifier.width(8.dp))
+    Icon(
+      imageVector = HedvigIcons.ArrowNorthEast,
+      contentDescription = EmptyContentDescription,
+      modifier = Modifier
+        .size(16.dp),
+    )
+  }
+}
+
+
 @HedvigPreview
 @Composable
-private fun PreviewAboutAppScreen() {
+private fun PreviewLinkRow() {
   HedvigTheme {
     Surface(color = HedvigTheme.colorScheme.backgroundPrimary) {
-      AboutAppScreen(
+      LinkRow(
+        link = "",
+        text = "Legal information",
+        onLinkClick = {},
+      )
+    }
+  }
+}
+
+@HedvigPreview
+@Composable
+private fun PreviewInformationScreen() {
+  HedvigTheme {
+    Surface(color = HedvigTheme.colorScheme.backgroundPrimary) {
+      InformationScreen(
         uiState = AboutAppUiState.Content(memberId = "123", deviceId = "123456"),
         onBackPressed = {},
         showOpenSourceLicenses = {},
@@ -281,7 +441,22 @@ private fun PreviewAboutAppScreen() {
         appVersionName = "11.3.2",
         appVersionCode = "43",
         navigateToNewConversation = {},
+        openUrl = {},
+        languageService = previewLanguageService,
       )
     }
+  }
+}
+
+private val previewLanguageService = object : LanguageService {
+  override fun setLanguage(language: Language) {}
+  override fun getSelectedLanguage(): Language {
+    return Language.EN_SE
+  }
+  override fun getLanguage(): Language {
+    return Language.EN_SE
+  }
+  override fun getLocale(): CommonLocale {
+    return CommonLocale.getDefault()
   }
 }
